[tests] TestFile: use int64_t to avoid warning about unsigned comparison

+++ xbmc/filesystem/test/TestFile.cpp
@@ -20,7 +20,7 @@ TEST(TestFile, Read)
   const size_t size = 1616;
   const size_t lines = 25;
   size_t addPerLine = newLine.length() - 1;
-  size_t realSize = size + lines * addPerLine;
+  int64_t realSize = size + lines * addPerLine;
 
   const std::string firstBuf  = "About" + newLine + "-----" + newLine + "XBMC is ";
   const std::string secondBuf = "an award-winning fre";
@@ -32,7 +32,7 @@ TEST(TestFile, Read)
   char buf[23];
   memset(buf, 0, sizeof(buf));
 
-  size_t currentPos;
+  int64_t currentPos;
   ASSERT_TRUE(file.Open(
     XBMC_REF_FILE_PATH("/xbmc/filesystem/test/reffile.txt")));
   EXPECT_EQ(0, file.GetPosition());
